api_atividade/requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
flask-httpauth-4.0.0
